+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Irrep.generators` and `Irreps.generators` functions to get the generators of the representations.
 
 ## [0.9.2] - 2022-09-29
 ### Fixed